Clone "Context" after every statement.
Makes Context cloneable and makes a clone of it after executing each
statement when not using the VM.

To do this, we make everything in Context cloneable and also hack the
StatementList to add a new run_and_track_state method that copies the
run method but tracks a vector of states and returns it.

We then plumb it back to the caller through the eval method on Context.